Implement search_sites
Allows specifying a list of `search_sites`, each of which will be searched
with each engine configured with `search_sites: True` by appending
`site:<site>` to the query. This effectively sets an allowlist of sites
from which we want to find results, allowing the administrator to
configure a custom search engine over a limited set of sites, similar to
Rollyo or Bing Custom Search.